requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1

# File Processing
python-multipart==0.0.6
python-magic==0.4.27
langchain-community==0.0.13
beautifulsoup4==4.12.2
pypdf2==3.0.1
python-docx==1.1.0
pdfminer.six==20231228

# LLM Integration
openai==1.3.7

# Caching
cachetools==5.3.2

# Validation & Serialization
pydantic==2.11.9
pydantic-settings==2.11.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
structlog==23.2.0
rich==13.7.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1